rename start to trace
