+++ cmd/cws-instrumentation/flags/flags_common.go
@@ -14,9 +14,4 @@ const (
 	Data = "data"
 	// SessionType represents the type of the user session that is being injected
 	SessionType = "session-type"
-
-	// SystemProbeGRPCAddr defines the system-probe GRPC addr
-	SystemProbeGRPCAddr = "grpc-addr"
-	// LogLevel defines the log level
-	LogLevel = "log-level"
 )
